Fix scoreboard bug
The scoreboard updated only if a won/loss granted one of the players a
point. This meant that, if a user played three rounds, and then tied on
their "new" first round of a game, the scoreboard, instead of updating
to "0 - 0", would still keep the old score.

I fixed this by adding the changeScore() function even when a tie
occured, so that the scoreboard would still reset.